fix12

+++ src/array_function.c
@@ -10,15 +10,15 @@ for(int i=0;i<size_src2;i++) {
 }
 
 void join_and_sort_int_arrays(int* src1, size_t size_src1, int* src2, size_t size_src2, int* dest) {
-    for(int i=0;i<size_src1;i++) {
+    for(size_t i=0;i<size_src1;i++) {
     dest[i] = src1[i];
 };
-for(int i=0;i<size_src2;i++) {
+for(size_t i=0;i<size_src2;i++) {
     dest[i+size_src1] = src2[i];
 };
 int n=0;
 while(n!=size_src1+size_src2 -1){
-    for(int i=0;i<size_src1+size_src2 -1;i++){
+    for(size_t i=0;i<size_src1+size_src2 -1;i++){
         if (dest[i]>dest[i+1]){
             int a=dest[i];
             dest[i]=dest[i+1];
